+++ lib/chef/knife/rackspace_server_delete.rb
@@ -66,7 +66,11 @@ class Chef
             msg_pair("Host ID", server.host_id)
             msg_pair("Name", server.name)
             msg_pair("Flavor", server.flavor.name)
-            msg_pair("Image", server.image.name)
+            if server.image
+              msg_pair("Image", server.image.name)
+            else
+              msg_pair("Image No image supplied booting from block volume")
+            end
             msg_pair("Public IP Address", ip_address(server, 'public'))
             msg_pair("Private IP Address", ip_address(server, 'private'))